# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      summary: Get top leaderboard entries
      parameters:
        - name: mode
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
    post:
      operationId: submitScore
      tags: [leaderboard]
      summary: Submit a score to the leaderboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitScoreBody"
      responses:
        "201":
          description: Score submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardEntry"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  schemas:
...
    LeaderboardEntry:
      type: object
      properties:
        id:
          type: integer
        playerName:
          type: string
        score:
          type: integer
        mode:
          type: string
        roster:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
      required:
        - id
        - playerName
        - score
        - mode
        - roster
        - createdAt
    SubmitScoreBody:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
        mode:
          type: string
        roster:
          type: object
          additionalProperties:
            type: string
      required:
        - score
        - mode
        - roster
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...