# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get top leaderboard entries
 */
export const getLeaderboardQueryLimitDefault = 50;

export const GetLeaderboardQueryParams = zod.object({
  "mode": zod.coerce.string().optional(),
  "limit": zod.coerce.number().default(getLeaderboardQueryLimitDefault)
})

export const GetLeaderboardResponseItem = zod.object({
  "id": zod.number(),
  "playerName": zod.string(),
  "score": zod.number(),
  "mode": zod.string(),
  "roster": zod.record(zod.string(), zod.string()),
  "createdAt": zod.string()
})
export const GetLeaderboardResponse = zod.array(GetLeaderboardResponseItem)


/**
 * @summary Submit a score to the leaderboard
 */
export const submitScoreBodyScoreMin = 0;



export const SubmitScoreBody = zod.object({
  "score": zod.number().min(submitScoreBodyScoreMin),
  "mode": zod.string(),
  "roster": zod.record(zod.string(), zod.string())
})


/**
 * @summary Get the currently authenticated user
 */
export const GetCurrentAuthUserHeader = zod.object({
  "Authorization": zod.string().optional().describe('Opaque session token — Bearer <sid>.')
})

export const GetCurrentAuthUserResponse = zod.object({
  "user": zod.union([zod.object({
  "id": zod.string(),
  "email": zod.string().email().nullable(),
  "firstName": zod.string().nullable(),
  "lastName": zod.string().nullable(),
  "profileImageUrl": zod.string().nullable()
}),zod.null()])
})


/**
 * @summary Start the browser OIDC login flow
 */
export const BeginBrowserLoginQueryParams = zod.object({
  "returnTo": zod.coerce.string().optional()
})


/**
 * @summary Complete the browser OIDC login flow
 */
export const HandleBrowserLoginCallbackQueryParams = zod.object({
  "code": zod.coerce.string().optional(),
  "state": zod.coerce.string().optional()
})


/**
 * @summary Clear the session and begin OIDC logout
 */
export const LogoutBrowserSessionHeader = zod.object({
  "Authorization": zod.string().optional().describe('Opaque session token — Bearer <sid>.')
})


/**
 * @summary Exchange a mobile OIDC code for a session token
 */







export const ExchangeMobileAuthorizationCodeBody = zod.object({
  "code": zod.string().min(1),
  "code_verifier": zod.string().min(1),
  "redirect_uri": zod.string().url().min(1),
  "state": zod.string().min(1),
  "nonce": zod.string().min(1).optional()
})

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  "token": zod.string()
})


/**
 * @summary Delete a mobile session token
 */
export const LogoutMobileSessionHeader = zod.object({
  "Authorization": zod.string().optional().describe('Opaque session token — Bearer <sid>.')
})

export const LogoutMobileSessionResponse = zod.object({
  "success": zod.boolean()
})