# lib/api-zod/src/generated/types/submitScoreBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SubmitScoreBodyRoster } from './submitScoreBodyRoster';

export interface SubmitScoreBody {
  /** @minimum 0 */
  score: number;
  mode: string;
  roster: SubmitScoreBodyRoster;
}